+++ integration-tests/environments/electron/app/mocha.js
@@ -37,6 +37,6 @@ return new Client({
     };
     // Add the integration test suite (in TypeScript)
     require("ts-node/register/transpile-only");
-    require("realm-integration-tests");
+    require("@realm/integration-tests");
   },
 });