+++ integration-tests/tests/src/index.ts
@@ -35,13 +35,17 @@ import { testSkipIf, suiteSkipIf } from "./utils/skip-if";
 global.describe.skipIf = suiteSkipIf;
 global.it.skipIf = testSkipIf;
 
-require("./tests/import-app-util");
-require("./tests/realm-constructor");
-require("./tests/serialization");
-require("./tests/objects");
-require("./tests/iterators");
-require("./tests/dynamic-schema-updates");
-require("./tests/bson");
-require("./tests/dictionary");
-require("./tests/credentials/anonymous");
-require("./tests/sync/mixed");
+// Using `require` instead of `import` here to ensure the Mocha globals (including `skipIf`) are set
+
+describe("Test Harness", () => {
+  require("./utils/import-app.test");
+});
+
+// Simplify once https://github.com/kraenhansen/mocha-remote/issues/58 gets solved
+describe.skipIf(environment.integration === false || environment.integration === "false", "Integration tests", () => {
+  require("./tests");
+});
+
+describe.skipIf(environment.performance !== true, "Performance tests", () => {
+  require("./performance-tests");
+});